+++ packages/compartment-mapper/NEWS.md
@@ -6,6 +6,8 @@ User-visible changes to the compartment mapper:
   of exit modules. Unlike import functions, the values of the exit module
   record are ignored.
   Any omitted exit module will cause an exception during archive creation.
+- Adds a `dev` option to archive and import workflows to include the
+  `devDependencies` of the entry package (but not other packages).
 - Fixes a missing file in the published assets for
   `@endo/compartment-mapper/node-powers.js`.